move text of slide 1+2 to bottom
