+++ src/components/Activities/Activity3/Game/GameWrapper.tsx
