+++ src/components/Activities/Activity3/Game/VariableCard.tsx
