+++ src/components/Activities/Activity3/index.tsx
@@ -6,19 +6,20 @@ import Carousel, { CarouselItemComponents } from '../../shared/Carousel';
 import ComputerRecs from './Animations/ComputerRecs';
 import ManyEmployees from './Animations/ManyEmployees';
 import PlantSprout from './Animations/PlantSprout';
-import { GameIntroSlide1, GameIntroSlide2 } from './GameIntroSlides';
-import GameWrapper from './GameWrapper/GameWrapper';
+import { GameIntroSlide1, GameIntroSlide2 } from './Game/GameIntroSlides';
+import GameWrapper from './Game/GameWrapper';
 
 function Activity3(): JSX.Element {
   const content: CarouselItemComponents[] = [
     {
       child: <ManyEmployees start={false} />,
-      topText: 'Youtube has around 10,000 employees',
-      bottomText: 'That\'s a lot of people--what are they all doing?',
+      bottomText: 'Youtube has around 10,000 employees',
+      bottomText2: 'That\'s a lot of people--what are they all doing?',
+      animationTime: 2,
     },
     {
       child: <ManyEmployees start={true} />,
-      topText: 'Different teams work on different ways to improve the product.',
+      bottomText2: 'Different teams work on different ways to improve the product.',
       animationTime: 2,
     },
     {
