+++ src/components/shared/Carousel.tsx
@@ -208,7 +208,7 @@ function Carousel(props: CarouselProps): JSX.Element {
 
                 {child.topText && <h2 id={'body-text'}> {child.topText} </h2>}
                 {child.child}
-                {child.bottomText && <h2 id={'body-text'}> {child.bottomText} </h2>}
+                {(child.bottomText || child.bottomText2) && <h2 id={'body-text'}> {child.bottomText} </h2>}
                 {child.bottomText2 && <h2 id={'body-text'}> {child.bottomText2} </h2>}
               </>
             }
