+++ src/components/styles/Activity3.scss
@@ -10,10 +10,10 @@ b {
 #many-employees {
   margin-left: 86px;
   width: 928px;
-  height: 386px;
+  height: 350px;
   position: absolute;
   background-repeat: no-repeat;
-  background-size: contain;
+  background-size: cover;
   background-image: url('../../assets/activity3/many_employees.svg');
 }
