+++ src/components/styles/Carousel.scss
@@ -18,6 +18,7 @@
   font: $font-title;
   font-size: 26px;
   margin: 5px;
+  height: 32px;
 }
 
 b {
